+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2023-06-17
+### Changed
+- Function signature of custom coercions to allow parsing more complex coercions types such as substr.
+
+### Added
+- Added new `_substr_[n:n]` COERCE identifier.
+
 ## [0.8.0] - 2023-06-10
 ### Added
 - Ability to register new, remove existing and replace existing COERCE types.
@@ -70,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial conversion from https://github.com/rust-playground/ksql.
 
-[Unreleased]: https://github.com/go-playground/ksql/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/go-playground/ksql/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/go-playground/ksql/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/go-playground/ksql/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/go-playground/ksql/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/go-playground/ksql/compare/v0.6.0...v0.6.1